File storage

+++ spec/toolkit/TestStorage.js
@@ -2,17 +2,30 @@ import fs from 'fs';
 import rimraf from 'rimraf';
 
 export default class TestStorage {
-    init() {
-        this.folder = fs.mkdtempSync(`${__dirname}/`);
+    init(bottle) {
+        this.folder = fs.mkdtempSync(`${__dirname}/temp_`);
+        afterEach(() => this.dispose());
+        if (bottle) {
+            this.patch(bottle);
+        }
+        return this;
     }
+
     getFolder() {
         if (!this.folder) {
             throw new Error('init not called yet');
         }
         return this.folder;
     }
+
     dispose() {
-        rimraf.sync(this.getFolder());
-        this.folder = null;
+        if (this.folder) {
+            rimraf.sync(this.getFolder());
+            this.folder = null;
+        }
+    }
+
+    patch(bottle) {
+        bottle.container.configuration.getStorageFolder = () => this.folder;
     }
 }
